P2/backend/requirements.txt
# pip==23.3.1
django==4.2
pillow==10.0
djangorestframework
django-filter
djangorestframework-simplejwt